bbocr_server/requirements.txt
aistudio-sdk==0.3.8
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
apsisocr==0.0.7
asgiref==3.9.2
asttokens==3.0.0
attrs==25.3.0
bce-python-sdk==0.9.46
beautifulsoup4==4.14.2
blinker==1.9.0
bnunicodenormalizer==0.1.7
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.1
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.118.0
fastdeploy-python==1.0.7
fastdeploy-tools==0.0.5
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2024.6.1
future==1.0.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
imagesize==1.4.1
iopath==0.1.10
ipykernel==6.30.1
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
layoutparser==0.3.4
line_profiler==5.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
matplotlib-inline==0.1.7
modelscope==1.30.0
mpmath==1.3.0
narwhals==2.6.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.2.6
onnxruntime==1.23.0
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.2.0
paddlepaddle==3.2.0
paddlex==3.2.1
pandas==2.3.3
parso==0.8.5
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
polars==1.33.1
portalocker==3.2.0
prettytable==3.16.0
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
pypdfium2==4.30.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
scipy==1.15.3
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.48.0
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
toml==0.10.2
tomli==2.2.1
torch==2.8.0+cpu
torchvision==0.23.0+cpu
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
ultralytics==8.3.204
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.16.0
watchdog==6.0.0
wcwidth==0.2.14
Werkzeug==3.1.3
wget==3.2
